BUGFIX: copy task should not crash if assets dir does not exist

+++ templates/__common__/config/gulp/copy.js
@@ -1,4 +1,5 @@
 // packages
+const chalk = require('chalk');
 const fs = require('fs-extra');
 
 // internal
@@ -6,8 +7,12 @@ const { isImagePath } = require('./utils');
 const paths = require('../paths');
 
 module.exports = async () => {
-  await fs.copy(paths.appAssets, paths.appDistAssets, {
-    dereference: true,
-    filter: src => !isImagePath(src),
-  });
+  if (await fs.pathExists(paths.appAssets)) {
+    await fs.copy(paths.appAssets, paths.appDistAssets, {
+      dereference: true,
+      filter: src => !isImagePath(src),
+    });
+  } else {
+    console.warn(chalk.grey`No "assets" directory found, skipping copy task.`);
+  }
 };
